Time slots fixed in ShiftReports, BatchGen and dailyops
Now the conversion Local2UTC is autonomous, also added ad hoc text messages for the different kind of passes (SDA4, DBA1 and MCM). Could create also other ones for the space passes (xTTC3, CDA) but since they are not nominal I don't think is necessary. 

Fixed the bug in the ShiftReports that created problems when uploading a Groundcon Uberlog export instead of  Spacon