Fixed BatchGen and ShiftReports
Fixed the bugs in the time selection for the "Other" option in both files. Fixed the visualization bug that displayed unicode characters in the ShiftReport comment section. Also now it creates an Excel and not a csv and the rows are ordered by Antenna, then by time. This is a good option only for ground, for space it should be ordered by Satellite and then by Time. I will update it asap